# swagger.yml
swagger: '2.0'
info:
  title: MiMQIP API
  description: Endpoints for the MiMQIP API
  version: "1.0.0"
host: api.mimqip
schemes:
  - https
basePath: /
produces:
  - application/json
paths:
  /rules:
    get:
      summary: Rules
      description: Endpoint returns a list of rules
      tags:
        - rule

      responses:
        200:
          description: Array of rules
          schema:
            type: array
            items:
              $ref: '#/definitions/Rule'
  /rules/{id}:
    get:
      summary: Rule by ID
      description: Access a single rule by ID
      tags:
        - rule
      parameters:
        - name: id
          in: path
          description: ID of rule to fetch
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: Rule
          schema:
            $ref: '#/definitions/Rule'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /patients:
      get:
        summary: Patients
        description: Endpoint returns a list of patients
        tags:
          - patient
        responses:
          200:
            description: Array of rules
            schema:
              type: array
              items:
                $ref: '#/definitions/Patient'
  /patients/{id}:
    get:
      summary: Patient by ID
      description: Access a single patient by ID
      tags:
        - patient
      parameters:
        - name: id
          in: path
          description: ID of rule to fetch
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: Rule
          schema:
            $ref: '#/definitions/Patient'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
  Rule:
    type: object
    properties:
      id:
        type: string
      priority:
        type: integer
        format: int32
      valid:
        type: boolean
      ruleType:
        type: string
        enum:
          - val1
          - val2
          - val3
          - val4
      name:
        type: string
  Patient:
    type: object
    properties:
      id:
        type: integer
        format: int64
      facilityId:
        type: integer
        format: int64
      valid:
        type: boolean
      startDate:
        type: string
        format: 'date-time'
      rcopiaLastUpdatedDate:
        type: string
        format: 'date-time'
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string